Create MAZE GENRATOR AND SOLVER
When running the program, the user needs to provide the number of rows and columns for the maze. The program then generates a random maze using Prim's algorithm and prints it. It also solves the maze using breadth-first search and prints the solution path.

Please note that the code provided here does not include any graphical representation of the maze. It focuses on the logic and algorithms involved in maze generation and solving.