Create genetic algorithm to solve the Traveling Salesman Problem (TSP)
This code uses classes to represent the City, Tour, and Population. It implements the main components of a genetic algorithm, including population initialization, fitness calculation, tournament selection, crossover, and mutation. The code also includes a helper function to generate a random set of cities.

To use this code, you need to provide the number of cities, the maximum X and Y coordinates for the cities, and the genetic algorithm parameters such as population size, number of generations, mutation rate, tournament size, and elitism. The program will then generate a random set of cities and apply the genetic algorithm to find the best tour solution for the Traveling Salesman Problem. Finally, it prints the best tour found.