Create  File Compression and Decompression Utility using C++ and the zlib library
This project demonstrates the implementation of file compression using the Deflate algorithm (a combination of Huffman coding and LZ77) and file decompression to restore the original files:
To use this code, you'll need to have the zlib library installed on your system and link it during compilation. The code provides a menu-driven interface to compress and decompress files. Users can choose to compress a file by providing the input file name and the output file name. Similarly, users can choose to decompress a file by providing the input file name and the output file name.